# docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handlers.NoteRequest:
    properties:
      content:
        type: string
      title:
        description: 标题最多100字
        maxLength: 100
        type: string
    required:
    - title
    type: object
  handlers.RegisterRequest:
    properties:
      password:
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: 这是一个基于 Go + Gin + MySQL 的简单记事本项目 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: 记事本 API
  version: "1.0"
paths:
  /api/login:
    post:
      consumes:
      - application/json
      description: 用户登录并获取Token
      parameters:
      - description: 用户名和密码
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                data:
                  properties:
                    expires_at:
                      type: integer
                    token:
                      type: string
                    user:
                      type: object
                  type: object
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      summary: 用户登录
      tags:
      - users
  /api/notes:
    post:
      consumes:
      - application/json
      description: 创建一条新笔记
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 笔记标题和内容
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/handlers.NoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      summary: 创建笔记
      tags:
      - notes
  /api/register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 用户名和密码
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/gin.H'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      summary: 用户注册
      tags:
      - users
schemes:
- http
swagger: "2.0"